Move readAll to buffer.ts
